This is another file that helped me practice testing in python
lexicon.py is takes a users input and checks to see if their input exists in one or more of the various lists. The lists are sorted into types of words that appear might appear in a game.  

After it has checked to see if the input exists, it returns a list of tuples in the format (TYPE, WORD). Pairing the word with its type. Also the program needs to accept input regardless of the case.